Add CSS selector for red-links in the `#p-personal` menu
Previously, when a user page didn't exist and appeared as a red links
in the `#p-personal` menu, it depended on styles from mediawiki core,
in `content.links.less#7` to have the color red. This style will be
removed from core soon, so a similar selector was added here to maintain
the red appearance.

Bug: T282884
Change-Id: Ib7521fe0062d067311c51f72480e92347de8632e